+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # SmartString Changelog
 
+## [Unreleased] - 2024-11-13
+
+- Added ::newArray() method to create a SmartArray of SmartStrings
+- Misc code and other minor improvements
+
 ## [1.3.2] - 2024-11-07
 
 - Fixed jsEscape() double escaping issue